requirements.txt
# Core FastAPI & Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
starlette==0.35.0

# Database & ORM
sqlalchemy==2.0.27
alembic==1.13.0
psycopg2-binary==2.9.9

# Data Validation & Serialization
pydantic==2.6.0
pydantic[email]==2.6.0

# Authentication & Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
PyJWT==2.9.0
bcrypt==4.1.2
cryptography==41.0.7

# Environment & Configuration
python-dotenv==1.0.0

# API Documentation
python-multipart==0.0.6



# Development & Testing (Optional)
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.25.2
